+++ umami/models/model_cads.py
@@ -1,6 +1,5 @@
 """Keras model of the CADS tagger."""
 from umami.configuration import logger  # isort:skip
-import json
 import os
 
 import h5py
@@ -12,7 +11,7 @@ import umami.tf_tools as utf
 import umami.train_tools as utt
 
 
-def cads_model(
+def create_cads_model(
     train_config: object,
     input_shape: tuple,
     continue_training: bool = False,
@@ -200,7 +199,7 @@ def cads_tagger(args, train_config, preprocess_config):
         )
 
     # Init CADS model
-    cads, epochs, init_epoch = cads_model(
+    cads_model, epochs, init_epoch = create_cads_model(
         train_config=train_config,
         input_shape=(metadata["n_trks"], metadata["n_trk_features"]),
         continue_training=train_config.continue_training,
@@ -282,11 +281,7 @@ def cads_tagger(args, train_config, preprocess_config):
         val_data_dict=val_data_dict,
         target_beff=working_point,
         frac_dict=eval_params["frac_values"],
-        dict_file_name=utt.get_validation_dict_name(
-            WP=working_point,
-            n_jets=n_jets_val,
-            dir_name=train_config.model_name,
-        ),
+        n_jets=n_jets_val,
         continue_training=train_config.continue_training,
     )
 
@@ -294,7 +289,7 @@ def cads_tagger(args, train_config, preprocess_config):
     callbacks.append(my_callback)
 
     logger.info("Start training")
-    history = cads.fit(
+    cads_model.fit(
         train_dataset,
         epochs=n_epochs,
         # TODO: Add a representative validation dataset for training (shown in stdout)
@@ -307,13 +302,3 @@ def cads_tagger(args, train_config, preprocess_config):
         workers=8,
         initial_epoch=init_epoch,
     )
-
-    # Dump dict into json
-    logger.info(f"Dumping history file to {train_config.model_name}/history.json")
-
-    # Make the history dict the same shape as the dict from the callbacks
-    hist_dict = utt.prepare_history_dict(history.history)
-
-    # Dump history file to json
-    with open(f"{train_config.model_name}/history.json", "w") as outfile:
-        json.dump(hist_dict, outfile, indent=4)